requirements.txt
streamlit>=1.22.0
requests>=2.28.2
flask>=2.2.3
flask-cors>=3.0.10
python-dotenv>=1.0.0
openai>=1.0.0
gunicorn>=20.1.0
pandas>=1.5.0
matplotlib>=3.5.0
numpy>=1.22.0 